move debug handle
